requirements.txt
fastapi
uvicorn[standard]
gspread
oauth2client
sentence-transformers
faiss-cpu
